My bubble work from 2/5/16-2/6/16
I worked on making circles. I started by making one blue circle and
then made it so that it changed its position on the page. I did this
using math.Random, learning about concatenating (+) things together
plus being able to make functions. I think the big influences in making
functions is knowing that a function is only made with things that you
define - similar to scratch. I didn’t realize that until monday of this
week. This week was the first week that I realized that variables are
defined based on the context and meaning that they acquire from the
functions that they are in (which are also constructed by me). After
making a blue circle that changed positions randomly on the page (I
first tried to use x and y coordinates like scratch because I saw in
the console that if you look at all of the attributes of a thing that x
and y are two of the attributes. But so are top and left. yesterday was
really the first time that the element.attribute notation in javascript
actually began to really mean anything and so when I realized that I
could use the element.style.(top/left) to change the thing and then
could make top and left a random number using math random via creation
of  function…alec and I made a circle that moved around the page every
time we refreshed. Also because we didn’t want it going off of the page
we subtracted 200px (the size of the circle) form the amount that it
could move on the page but that only worked on the bottom not on the
top). Anyway, then we worked on this function called set interval which
is like a timer. It makes things move on their own where you do
setInternval(function,time in milliseconds between the function being
run and the next time) and then once we did this, the circle would love
around the screen on a loop.

Then we worked on changing the colors of the thing via the “r, g, b” -
(note all that javascript does is modify html and css and that’s what
is reported in the console. It is a dynamic thing that changes the html
and css and the console capture that at every point. think of the
console as something that “just” sees the images in every frame of a
movie as opposed to being able to see the movie running.)

Also, aside, alec thought up a GREAT analogy. That a variable is an
address on a house (before he used to call it a label on a box). There
are things inside of the house. When you make a function you (you have
a bird’s eye view of the village) someone goes into the house, does
something (the function) and then if the function is returning a thing
(denoted by the “()” at the end of a function eg: function ()) then you
can return that value and once you do that the value can be assigned to
a thing (an element, a style, a color. If the value is not assigned to
a color or a thing, just think of a function kind of like a thing that
you made “just” to make strings, or numbers.

For loops don’t return anything. They are like the idea of an “era era”
on a record. They aren’t actually the record itself and don’t actually
return anything because for loops aren’t functions. You can set for
loops to be inside of functions and then you can use the return from
THOSE functions to do stuff but that’s only once you make the function
into a thing.

After doing set interval, did it with colors, and then with sizes of
the circles. Then made the sizes of the circle respond to click (using
add event listener) and then made a “toggle” function that ran when
click and then turned off when clicked again. This is a bit more
complicated and I’m not sure if I fully get this yet - need to look at
it again but can’t summarize right now.